添加ViewPager的循环滚动以及利用clipToPadding属性解决触及屏幕边缘时不能滑动问题
