+++ app/src/main/AndroidManifest.xml
@@ -16,6 +16,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity android:name=".CircleViewPagerActivity"/>
+        <activity android:name=".StandardViewPagerActivity"/>
     </application>
 
 </manifest>
