+++ app/src/main/res/layout/activity_circle_viewpager.xml
@@ -1,0 +1,22 @@
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="160dp"
+        android:layout_centerInParent="true"
+        android:background="#aadc71ff">
+
+        <android.support.v4.view.ViewPager
+            android:id="@+id/id_viewpager"
+            android:layout_width="match_parent"
+            android:layout_height="120dp"
+            android:layout_gravity="center"
+            android:clipToPadding="false"
+            android:paddingLeft="60dp"
+            android:paddingRight="60dp"/>
+
+    </FrameLayout>
+
+</RelativeLayout>
